Instead of creating new Activities on top when touching a notification, replace the current one.
Scenario: open ChatWindow, but moved to the background, by launcher or
other app. Incoming msg creates notification. Opening notification
creates a new ChatWindow on top of the existing one, creating a stack
of multiple ChatWindow activities.

The included patch removes the old ChatWindow activity in favor for the
one opened by the notification.