Add support for receiving vCard-based avatar images
Evaluate XEP-0153 tags and store avatar images and their hashes in the
local roster db. The frontend will be able to pick them up this way and
display them.

The roster db is used as a cache for the images to avoid unneeded
traffic. Only if the cached data and a newly provided hash mismatch, the
image is retrieved from the server.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>