Prevent notifications for open chats with user activity
When a message arrives for an open chat, do not immediately deliver it
but defer it by 10 s. If the user interacts with the chat in that
period, it is safe to assume it also recognized the message. In that
case we can cancel the notification, otherwise we deliver it as usual
once the timer fires.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>